feat(examples): shared scenario runner + SaaS startup showcase (episode 2)
Extract examples/_scenario/runner.py (+ reset.py) — a parameterized end-to-end
runner that provisions a graph and loads a whole company from a Scenario plus
its agents, mappings, and policy docs. Coffee's main.py becomes a ~50-line thin
wrapper; its _reset.py/download_bundles.py fold into the shared runner. A new
'slug' field on Scenario drives demo slot / credentials / tags / bundle names,
and staggered_prepaid now rotates cohorts across the customer roster.

Episode 2 — examples/saas_startup_demo (Cadence Labs): a seed-funded B2B SaaS
that proves the thesis. It's a new data.py/agents.py/mappings.py/policies.py +
thin main.py — no runner fork. A structurally different reveal renders on the
filed report: ~79% gross margin but an operating LOSS of $781k (R&D $535k +
S&M $506k each their own IS line), $2.36M cash of which $1.10M is deferred
revenue, and the cash-flow statement shows the prepayment float (+$380k) funding
the burn — the deferred-revenue runway illusion. SHACL conforms, Arelle valid.
