+++ examples/saas_startup_demo/output/.gitignore
@@ -1,0 +1,3 @@
+# Runtime demo output — regenerated on each run (contains run-specific graph/report IDs).
+*
+!.gitignore
